Create README.md
Write a program to simulate an ATM machine. 

The system prompts the user to enter a pin number before display the menu, assume Table 1 is a list of ten users pin number. 



Please enter your pin number: 


If the pin number is incorrect, ask the user to enter the correct pin number. 

Table 1: Pin Number of 10 users.

0 	111111 
1	222222
2	333333
3	444444
4	555555
5	666666
6	777777
7	888888
8	999999
9	123456


Once the pin number is accepted, the main menu is display as shown below. 

 			1 - View the current balance
	2 – Withdraw money
	3 – Deposit money
	4 – Exit the main menu
			5 – Logout from the system
	Enter your choice:

Table 2 shows list of current balance in the 10 accounts customers that match with the pin number in Table 1.

Table 2: Current Balance of 10 users.
0 	50000.00
1	10000.00
2	8900.00
3	4750.00
4	14000.00
5	87000.00
6	930.00
7	7500.00
8	2740.00
9	3480.00

If the user tries to withdraw more money than is available, the program reports a “Not enough balance” message. Once the user select choice number 4, the system will go back to the pin number menu. To stop the system, press 5.
Sample output:
********************* WELCOME TO MY ATM *********************

Please enter your pin number: 333333


******************** MAIN MENU ************************

1: check balance
2: withdraw
3: deposit
4: exit
5: logout

Enter a choice: 1

The balance is RM8900.00

******************** MAIN MENU ************************

1: check balance
2: withdraw
3: deposit
4: exit
5: logout
Enter a choice: 2

******************** WITHDRAW MENU ************************

Enter amount to withdraw: 1000.00

Your new balance is: 7900.00




******************** MAIN MENU ************************

1: check balance
2: withdraw
3: deposit
4: exit
5: logout
Enter a choice: 4


********************* WELCOME TO MY ATM *********************

Please enter your pin number: